Add content pages support for 200/404
