+++ VirtoCommerce.Storefront.Model/ISeoInfoService.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using VirtoCommerce.Storefront.Model.StaticContent;
 using VirtoCommerce.Storefront.Model.Stores;
 
 namespace VirtoCommerce.Storefront.Model
@@ -8,5 +9,7 @@ namespace VirtoCommerce.Storefront.Model
         Task<SeoInfo[]> GetSeoInfosBySlug(string slug);
 
         Task<SeoInfo[]> GetBestMatchingSeoInfos(string slug, Store store, string currentCulture);
+
+        ContentItem GetContentItem(string slug, string culture);
     }
 }
