+++ VirtoCommerce.Storefront/Controllers/Api/ApiCommonController.cs
@@ -80,24 +80,7 @@ namespace VirtoCommerce.Storefront.Controllers.Api
 
             if (result.EntityInfo == null)
             {
-                var pageUrl = slug == "__index__home__page__" ? "/" : $"/{slug}";
-                try
-                {
-                    var pages = WorkContext.Pages.Where(p =>
-                        string.Equals(p.Url, pageUrl, StringComparison.OrdinalIgnoreCase)
-                        || string.Equals(p.Url, slug, StringComparison.OrdinalIgnoreCase)
-                    );
-
-                    var page = pages.FirstOrDefault(x => x.Language.CultureName.EqualsInvariant(culture))
-                                            ?? pages.FirstOrDefault(x => x.Language.IsInvariant)
-                                            ?? pages.FirstOrDefault(x => x.AliasesUrls.Contains(pageUrl, StringComparer.OrdinalIgnoreCase));
-                    result.ContentItem = page;
-
-                }
-                catch
-                {
-                    //do nothing
-                }
+                result.ContentItem = _seoInfoService.GetContentItem(slug, culture);
             }
 
             return result;
