+++ VirtoCommerce.Storefront/Startup.cs
@@ -93,7 +93,7 @@ namespace VirtoCommerce.Storefront
             services.AddSingleton<ICurrencyService, CurrencyService>();
             services.AddSingleton<ISlugRouteService, SlugRouteService>();
             services.AddSingleton<IMemberService, MemberService>();
-            services.AddSingleton<ISeoInfoService, SeoInfoServise>();
+            services.AddSingleton<ISeoInfoService, SeoInfoService>();
             services.AddSingleton<ISpaRoutesService, SpaRoutesService>();
 
             services.AddSingleton<IStaticContentService, StaticContentService>();
